refactor: extract MatchNotificationCoordinator and MatchUiState from MatchViewModel
MatchViewModel was 870 lines. Extracted:
- minuteOfPlay() and fireNotification() logic → MatchNotificationCoordinator
- PlayerTimeItem, MatchUiState, SubstitutionItem, EndPeriodState → MatchUiState.kt

MatchViewModel is now under 800 lines.
