+++ viewmodel/src/commonMain/kotlin/com/jesuslcorominas/teamflowmanager/viewmodel/MatchViewModel.kt
@@ -6,14 +6,8 @@ import com.jesuslcorominas.teamflowmanager.domain.analytics.AnalyticsEvent
 import com.jesuslcorominas.teamflowmanager.domain.analytics.AnalyticsParam
 import com.jesuslcorominas.teamflowmanager.domain.analytics.AnalyticsTracker
 import com.jesuslcorominas.teamflowmanager.domain.analytics.CrashReporter
-import com.jesuslcorominas.teamflowmanager.domain.model.Match
 import com.jesuslcorominas.teamflowmanager.domain.model.MatchStatus
-import com.jesuslcorominas.teamflowmanager.domain.model.Player
-import com.jesuslcorominas.teamflowmanager.domain.model.PlayerActivityInterval
-import com.jesuslcorominas.teamflowmanager.domain.model.PlayerTime
 import com.jesuslcorominas.teamflowmanager.domain.model.PlayerTimeStatus
-import com.jesuslcorominas.teamflowmanager.domain.model.ScorePoint
-import com.jesuslcorominas.teamflowmanager.domain.model.TimelineEvent
 import com.jesuslcorominas.teamflowmanager.domain.usecase.EndTimeoutUseCase
 import com.jesuslcorominas.teamflowmanager.domain.usecase.ExportMatchReportToPdfUseCase
 import com.jesuslcorominas.teamflowmanager.domain.usecase.FinishMatchUseCase
@@ -75,6 +69,7 @@ class MatchViewModel(
     private val getTeamUseCase: GetTeamUseCase,
 ) : ViewModel() {
     private val teamFlow = getTeamUseCase().stateIn(viewModelScope, SharingStarted.Eagerly, null)
+    private val notificationCoordinator = MatchNotificationCoordinator(notifyPresidentMatchEvent)
 
     private val _uiState = MutableStateFlow<MatchUiState>(MatchUiState.Loading)
     val uiState: StateFlow<MatchUiState> = _uiState.asStateFlow()
@@ -132,7 +127,11 @@ class MatchViewModel(
                         startPlayerTimersBatchUseCase(it.id, it.startingLineupIds, currentTime)
                     }
 
-                    fireNotification { MatchEventNotification.Start(it.teamName, it.opponent) }
+                    notificationCoordinator.fireNotification(
+                        scope = viewModelScope,
+                        team = teamFlow.value,
+                        matchId = it.id.toString(),
+                    ) { MatchEventNotification.Start(it.teamName, it.opponent) }
                 }
             }
         }
@@ -182,7 +181,11 @@ class MatchViewModel(
                     )
 
                     val finishedMatch = currentState.match
-                    fireNotification {
+                    notificationCoordinator.fireNotification(
+                        scope = viewModelScope,
+                        team = teamFlow.value,
+                        matchId = finishedMatch.id.toString(),
+                    ) {
                         MatchEventNotification.End(finishedMatch.teamName, finishedMatch.opponent, finishedMatch.goals, finishedMatch.opponentGoals)
                     }
                 }
@@ -521,9 +524,13 @@ class MatchViewModel(
 
                     val goalMatchId = currentState.match.id
                     val snapshotTime = _currentTime.value
-                    fireNotification {
+                    notificationCoordinator.fireNotification(
+                        scope = viewModelScope,
+                        team = teamFlow.value,
+                        matchId = goalMatchId.toString(),
+                    ) {
                         val updatedMatch = getMatchById(goalMatchId).first() ?: return@fireNotification null
-                        val minuteOfPlay = minuteOfPlay(updatedMatch, snapshotTime)
+                        val minuteOfPlay = notificationCoordinator.minuteOfPlay(updatedMatch, snapshotTime)
                         MatchEventNotification.Goal(
                             teamName = updatedMatch.teamName,
                             opponentName = updatedMatch.opponent,
@@ -576,9 +583,13 @@ class MatchViewModel(
 
                     val goalMatchId = currentState.match.id
                     val snapshotTime = _currentTime.value
-                    fireNotification {
+                    notificationCoordinator.fireNotification(
+                        scope = viewModelScope,
+                        team = teamFlow.value,
+                        matchId = goalMatchId.toString(),
+                    ) {
                         val updatedMatch = getMatchById(goalMatchId).first() ?: return@fireNotification null
-                        val minuteOfPlay = minuteOfPlay(updatedMatch, snapshotTime)
+                        val minuteOfPlay = notificationCoordinator.minuteOfPlay(updatedMatch, snapshotTime)
                         MatchEventNotification.Goal(
                             teamName = updatedMatch.teamName,
                             opponentName = updatedMatch.opponent,
@@ -595,41 +606,6 @@ class MatchViewModel(
                 crashReporter.recordException(e)
                 crashReporter.log("Error registering opponent goal: ${e.message}")
                 throw e
-            }
-        }
-    }
-
-    private fun minuteOfPlay(
-        match: Match,
-        currentTimeMs: Long,
-    ): String? {
-        val currentPeriod =
-            match.periods.firstOrNull { it.startTimeMillis > 0L && it.endTimeMillis == 0L }
-                ?: return null
-        val periodDurationMs = currentPeriod.periodDuration
-        val periodDurationMin = (periodDurationMs / 60_000L).toInt()
-        val completedMin = (currentPeriod.periodNumber - 1) * periodDurationMin
-        val elapsedInPeriodMs = (currentTimeMs - currentPeriod.startTimeMillis).coerceAtLeast(0L)
-        val elapsedInPeriodMin = (elapsedInPeriodMs / 60_000L).toInt()
-        return if (elapsedInPeriodMs <= periodDurationMs) {
-            (completedMin + elapsedInPeriodMin).toString()
-        } else {
-            val extraMin = elapsedInPeriodMin - periodDurationMin
-            "${completedMin + periodDurationMin}+$extraMin"
-        }
-    }
-
-    private fun fireNotification(buildEvent: suspend () -> MatchEventNotification?) {
-        viewModelScope.launch {
-            runCatching {
-                val team = teamFlow.value ?: return@runCatching
-                val teamRemoteId = team.remoteId ?: return@runCatching
-                val clubRemoteId = team.clubRemoteId ?: return@runCatching
-                val matchId =
-                    (_uiState.value as? MatchUiState.Success)?.match?.id?.toString()
-                        ?: return@runCatching
-                val event = buildEvent() ?: return@runCatching
-                notifyPresidentMatchEvent(event, matchId, teamRemoteId, clubRemoteId)
             }
         }
     }
@@ -739,51 +715,11 @@ class MatchViewModel(
         }
     }
 
-    private fun List<Player>.toPlayerItems(
-        playerTimes: List<PlayerTime>,
-        currentTime: Long,
-        captainId: Long,
-    ): List<PlayerTimeItem> =
-        this.map { player ->
-            val playerTime = playerTimes.find { it.playerId == player.id }
-            val displayTime =
-                if (playerTime != null) {
-                    calculateCurrentTime(
-                        playerTime.elapsedTimeMillis,
-                        playerTime.isRunning,
-                        playerTime.lastStartTimeMillis,
-                        currentTime,
-                    )
-                } else {
-                    0L
-                }
-            PlayerTimeItem(
-                player = player,
-                timeMillis = displayTime,
-                isRunning = playerTime?.isRunning ?: false,
-                isPaused = playerTime?.status == PlayerTimeStatus.PAUSED,
-                isCaptain = player.id == captainId,
-            )
-        }
-
     private fun observeTime() {
         viewModelScope.launch {
             timeTicker.timeFlow.collect { now ->
                 _currentTime.value = now
             }
-        }
-    }
-
-    private fun calculateCurrentTime(
-        elapsedTimeMillis: Long,
-        isRunning: Boolean,
-        lastStartTimeMillis: Long?,
-        currentTimeMillis: Long,
-    ): Long {
-        return if (isRunning && lastStartTimeMillis != null) {
-            (elapsedTimeMillis + (currentTimeMillis - lastStartTimeMillis)).coerceAtLeast(0L)
-        } else {
-            elapsedTimeMillis.coerceAtLeast(0L)
         }
     }
 
@@ -826,45 +762,3 @@ class MatchViewModel(
 
     companion object
 }
-
-data class PlayerTimeItem(
-    val player: Player,
-    val timeMillis: Long,
-    val isRunning: Boolean,
-    val isPaused: Boolean,
-    val substitutionCount: Int = 0,
-    val isCaptain: Boolean = false,
-)
-
-sealed class MatchUiState {
-    data object Loading : MatchUiState()
-
-    data object NoMatch : MatchUiState()
-
-    data class Success(
-        val match: Match,
-        val currentTime: Long,
-        val playerTimes: List<PlayerTimeItem>,
-        val timelineEvents: List<TimelineEvent> = emptyList(),
-    ) : MatchUiState()
-
-    data class Finished(
-        val match: Match,
-        val currentTime: Long,
-        val playerTimes: List<PlayerTimeItem>,
-        val substitutions: List<SubstitutionItem>,
-        val timelineEvents: List<TimelineEvent> = emptyList(),
-        val scoreEvolution: List<ScorePoint> = emptyList(),
-        val playerActivity: List<PlayerActivityInterval> = emptyList(),
-    ) : MatchUiState()
-}
-
-data class SubstitutionItem(
-    val playerOut: Player,
-    val playerIn: Player,
-    val matchElapsedTimeMillis: Long,
-)
-
-data class EndPeriodState(
-    val isBreak: Boolean,
-)
